♻️ [devext] move button to end the session next to the session info
